% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{create_module_app}
\alias{create_module_app}
\alias{run_module_app}
\title{Create or run a simple shiny app given a module name}
\usage{
create_module_app(module, id = "m", ui_args = list(), server_args = list())

run_module_app(module, id = "m", ui_args = list(), server_args = list())
}
\arguments{
\item{module}{Character. The module name in snake case.}

\item{id}{Character. Module's id, "m" by default as there is no need to change it.}

\item{ui_args}{Named list. Arguments to be passed to the UI module's function.}

\item{server_args}{Named list. Arguments to be passed to the Server module's function.}
}
\value{
The app shiny object.

NULL This function is called for its side effects.
}
\description{
The purpose of this function is to be able to examine and test any Shiny module on this
packages without having to develop an actual shiny app.
}
\examples{
create_module_app("plotDistOneline",server_args =  list(x = shiny::reactive(rnorm(100))))


}
