% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{get_csdta_colours}
\alias{get_csdta_colours}
\title{Get the csdta corporative colours}
\usage{
get_csdta_colours(colours = NULL, show = FALSE)
}
\arguments{
\item{colours}{character; a character vector of colour names to get. If not provided, all colours are returned.}

\item{show}{logical; if \code{TRUE}, the colours are displayed instead of returned.}
}
\value{
Named string vector with csdta colours.
}
\description{
Get the csdta corporative colours
}
