% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_oneline.R
\name{plot_dist_oneline}
\alias{plot_dist_oneline}
\title{Plot a one-dimensional distribition in a narrow space}
\usage{
plot_dist_oneline(x, limits = NULL)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{limits}{Limits of the plot passed to \code{scale_x_continuous()}.}
}
\description{
The aim of this plot is to be able to picture the distribution of a one-dimensional set of values
using the least vertical space possible.
}
\details{
Kernel density and interpolation is used to weight every point to be able to use the weighted
value as an aesthetic that helps enhancing which regions are most likely.

This function is expected to be used repeatedly with different samples, and therefore, to be
able to visually see the changes in distribution it's desirable that the plot has the same limits
over all samples. This behaviour however is beyond the scope of the function because it has no memory,
it's stateless. Therefore, any complex behaviour has to be controlled from outside the function.
}
\examples{
plot_dist_oneline(rexp(100,2))
plot_dist_oneline(rnorm(10000),limits = c(-4,4))
plot_dist_oneline(runif(20,0,1))
plot_dist_oneline(rpois(200,1))
plot_dist_oneline(rbinom(200,10,.5))

}
