% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{show_csdta_colours}
\alias{show_csdta_colours}
\title{Show csdta colors}
\usage{
show_csdta_colours(colours = NULL)
}
\arguments{
\item{colours}{a subseting vector to filter colours. If \code{NULL} provided, all colours are selected.}
}
\description{
Helper function that calls \code{get_csdta_colours()} with \code{show = TRUE} so it doesn return anything.
}
\examples{
show_csdta_colours()
show_csdta_colours(c(1:3))
show_csdta_colours(c("red","bright_green"))

}
