% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shycs-package.R
\docType{package}
\name{shycs-package}
\alias{shycs}
\alias{shycs-package}
\title{shycs: Modules and themes for csdta Shiny apps.}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Xavier Farré Manzorro \email{xaviescacs@gmail.com}

Authors:
\itemize{
  \item Xavier Farré Manzorro \email{xaviescacs@gmail.com}
}

}
\keyword{internal}
