% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_header.R
\name{ui_csdta_header}
\alias{ui_csdta_header}
\title{Get the header for a csdta shiny app}
\usage{
ui_csdta_header(
  title,
  theory = TRUE,
  credits = TRUE,
  languages,
  csdta_link = "/",
  github_link = "/"
)
}
\arguments{
\item{title}{Character of length 1. Title of the app.}

\item{theory}{Boolean. Indicates whether or not to show the theory button.}

\item{credits}{Boolean. Indicates whether or not to show the credits button.}

\item{languages}{Character. Languages to be displayed, the sorting will be respcted.}

\item{csdta_link}{Character of length 1. href of the csdta logo anchor.}

\item{github_link}{Character of length 1. href of the github logo anchor.}
}
\value{
The header shyni.tag object
}
\description{
Get the header for a csdta shiny app
}
\examples{

ui_csdta_header(languages = "ca",title = "This is an app title")
}
