% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_dist_oneline}
\alias{theme_dist_oneline}
\title{Title}
\usage{
theme_dist_oneline(...)
}
\arguments{
\item{...}{Params to be passed to \code{theme_grey()}.}
}
\value{
A ggplot2 theme designed to plot a distribution over the X axis using
}
\description{
Title
}
